+++ panweave/src/lib.rs
@@ -60,6 +60,8 @@ pub use panweave_smart_energy as smart_energy;
 #[cfg(feature = "smart-energy")]
 pub mod cbke;
 #[cfg(feature = "smart-energy")]
+pub mod smart_energy_drivers;
+#[cfg(feature = "smart-energy")]
 pub mod smart_energy_endpoints;
 
 use heapless::Deque;